+++ src/test/kotlin/no/nav/helsearbeidsgiver/integrasjonstest/ApplicationTest.kt
@@ -7,25 +7,15 @@ import io.ktor.http.HttpStatusCode
 import kotlinx.coroutines.runBlocking
 import no.nav.helsearbeidsgiver.Producer
 import no.nav.helsearbeidsgiver.forespoersel.Forespoersel
-import no.nav.helsearbeidsgiver.forespoersel.ForespoerselEntitet
-import no.nav.helsearbeidsgiver.forespoersel.Status
 import no.nav.helsearbeidsgiver.innsending.InnsendingStatus
 import no.nav.helsearbeidsgiver.inntektsmelding.InntektsmeldingResponse
-import no.nav.helsearbeidsgiver.kafka.forespoersel.pri.PriMessage
 import no.nav.helsearbeidsgiver.soeknad.Sykepengesoeknad
 import no.nav.helsearbeidsgiver.testcontainer.LpsApiIntegrasjontest
 import no.nav.helsearbeidsgiver.utils.TestData
-import no.nav.helsearbeidsgiver.utils.buildForespoerselMottattJson
-import no.nav.helsearbeidsgiver.utils.buildForespoerselOppdatertJson
-import no.nav.helsearbeidsgiver.utils.buildForspoerselBesvartMelding
 import no.nav.helsearbeidsgiver.utils.buildJournalfoertInntektsmelding
 import no.nav.helsearbeidsgiver.utils.gyldigSystembrukerAuthToken
-import no.nav.helsearbeidsgiver.utils.jsonMapper
 import no.nav.helsearbeidsgiver.utils.wrapper.Orgnr
 import org.apache.kafka.clients.producer.ProducerRecord
-import org.jetbrains.exposed.sql.and
-import org.jetbrains.exposed.sql.selectAll
-import org.jetbrains.exposed.sql.transactions.transaction
 import org.junit.jupiter.api.Test
 import java.util.UUID
 
@@ -128,249 +118,5 @@ class ApplicationTest : LpsApiIntegrasjontest() {
         val soeknadListe =
             repositories.soeknadRepository.hentSoeknaderMedVedtaksperiodeId(vedtaksperiodeId).map { it.id }
         soeknadListe shouldBe listOf(soeknadId)
-    }
-
-    @Test
-    fun `leser oppdatert forespoersel fra kafka og henter det via api`() {
-        val forespoerselId = UUID.randomUUID()
-        val oppdatertForespoerselId = UUID.randomUUID()
-
-        val forespoerselMottattJson = buildForespoerselMottattJson(forespoerselId = forespoerselId)
-
-        val forespoerselOppdaterJson =
-            buildForespoerselOppdatertJson(
-                forespoerselId = oppdatertForespoerselId,
-                eksponertForespoerselId = forespoerselId,
-            )
-
-        val priRecord = ProducerRecord(priTopic, "key", forespoerselMottattJson)
-        val oppdatertPriRecord = ProducerRecord(priTopic, "key", forespoerselOppdaterJson)
-        Producer.sendMelding(priRecord)
-        Producer.sendMelding(oppdatertPriRecord)
-
-        runBlocking {
-            val responseOppdatertFsp =
-                fetchWithRetry(
-                    url = "http://localhost:8080/v1/forespoersel/$oppdatertForespoerselId",
-                    token = mockOAuth2Server.gyldigSystembrukerAuthToken("810007842"),
-                )
-            val oppdatertFsp = responseOppdatertFsp.body<Forespoersel>()
-            val responseFsp =
-                fetchWithRetry(
-                    url = "http://localhost:8080/v1/forespoersel/$forespoerselId",
-                    token = mockOAuth2Server.gyldigSystembrukerAuthToken("810007842"),
-                )
-            val forespoerselSvar = responseFsp.body<Forespoersel>()
-
-            forespoerselSvar.navReferanseId shouldBe forespoerselId
-            oppdatertFsp.navReferanseId shouldBe oppdatertForespoerselId
-            oppdatertFsp.status shouldBe Status.AKTIV
-            forespoerselSvar.status shouldBe Status.FORKASTET
-
-            transaction(db) {
-                ForespoerselEntitet
-                    .selectAll()
-                    .where {
-                        (ForespoerselEntitet.navReferanseId eq oppdatertForespoerselId) and
-                            (ForespoerselEntitet.eksponertForespoerselId eq forespoerselId)
-                    }.count() shouldBe 1
-            }
-        }
-    }
-
-    @Test
-    fun `leser oppdatert forespoersel når den ikke finnes i db`() {
-        val forespoerselId = UUID.randomUUID()
-        val oppdatertForespoerselId = UUID.randomUUID()
-        val vedtaksperiodeId = UUID.randomUUID()
-
-        val forespoerselOppdaterJson =
-            buildForespoerselOppdatertJson(
-                forespoerselId = oppdatertForespoerselId,
-                eksponertForespoerselId = forespoerselId,
-                vedtaksperiodeId = vedtaksperiodeId,
-            )
-
-        val oppdatertPriRecord = ProducerRecord(priTopic, "key", forespoerselOppdaterJson)
-
-        Producer.sendMelding(oppdatertPriRecord)
-
-        sjekkOmForespoerselFinnesIDB(oppdatertForespoerselId, forespoerselId, vedtaksperiodeId)
-    }
-
-    @Test
-    fun `leser oppdatert forespoersel når eksponert forespoersel er besvart`() {
-        val forespoerselId = UUID.randomUUID()
-        val oppdatertForespoerselId = UUID.randomUUID()
-        val vedtaksperiodeId = UUID.randomUUID()
-
-        val forespoerselMottattJson = buildForespoerselMottattJson(forespoerselId = forespoerselId, vedtaksperiodeId)
-        val priMessage = jsonMapper.decodeFromString<PriMessage>(forespoerselMottattJson)
-        val forespoersel = priMessage.forespoersel
-        if (forespoersel != null) {
-            services.forespoerselService.lagreNyForespoersel(
-                forespoersel = forespoersel,
-            )
-            services.forespoerselService.settBesvart(forespoerselId)
-        }
-
-        val forespoerselOppdaterJson =
-            buildForespoerselOppdatertJson(
-                forespoerselId = oppdatertForespoerselId,
-                eksponertForespoerselId = forespoerselId,
-                vedtaksperiodeId = vedtaksperiodeId,
-            )
-
-        val oppdatertPriRecord = ProducerRecord(priTopic, "key", forespoerselOppdaterJson)
-
-        Producer.sendMelding(oppdatertPriRecord)
-
-        runBlocking {
-            val responseOppdatertFsp =
-                fetchWithRetry(
-                    url = "http://localhost:8080/v1/forespoersel/$oppdatertForespoerselId",
-                    token = mockOAuth2Server.gyldigSystembrukerAuthToken("810007842"),
-                )
-            val oppdatertFsp = responseOppdatertFsp.body<Forespoersel>()
-            oppdatertFsp.navReferanseId shouldBe oppdatertForespoerselId
-            oppdatertFsp.status shouldBe Status.AKTIV
-
-            transaction(db) {
-                ForespoerselEntitet
-                    .selectAll()
-                    .where {
-                        (ForespoerselEntitet.navReferanseId eq oppdatertForespoerselId) and
-                            (ForespoerselEntitet.eksponertForespoerselId eq forespoerselId)
-                    }.count() shouldBe 1
-            }
-            val responseFsp =
-                fetchWithRetry(
-                    url = "http://localhost:8080/v1/forespoersel/$forespoerselId",
-                    token = mockOAuth2Server.gyldigSystembrukerAuthToken("810007842"),
-                )
-            val forespoerselSvar = responseFsp.body<Forespoersel>()
-            forespoerselSvar.navReferanseId shouldBe forespoerselId
-            forespoerselSvar.status shouldBe Status.BESVART
-        }
-    }
-
-    @Test
-    fun `Avviser duplikat forespoersel`() {
-        val oppdatertForespoerselId = UUID.randomUUID()
-        val eksponertForespoerselId = UUID.randomUUID()
-        val vedtaksperiodeId = UUID.randomUUID()
-        val forespoerselMottattJson =
-            buildForespoerselOppdatertJson(
-                forespoerselId = oppdatertForespoerselId,
-                eksponertForespoerselId = eksponertForespoerselId,
-                vedtaksperiodeId = vedtaksperiodeId,
-            )
-
-        // Sender forespoersel til Kafka for første gang
-        val priRecord = ProducerRecord(priTopic, "key", forespoerselMottattJson)
-        Producer.sendMelding(priRecord)
-        // Henter forespoersel fra db
-        sjekkOmForespoerselFinnesIDB(oppdatertForespoerselId, eksponertForespoerselId, vedtaksperiodeId)
-        // Sender samme forespoersel til Kafka på nytt
-        Producer.sendMelding(priRecord)
-
-        sjekkOmDetFinnesKunEnForespoerselIDB(oppdatertForespoerselId)
-    }
-
-    @Test
-    fun `Setter forespoersel som besvart`() {
-        val forespoerselId = UUID.randomUUID()
-        val oppdatertForespoerselId = UUID.randomUUID()
-        val vedtaksperiodeId = UUID.randomUUID()
-
-        sendForespoerselMottatt(forespoerselId, vedtaksperiodeId)
-        sendForespoerselOppdatert(oppdatertForespoerselId, forespoerselId, vedtaksperiodeId)
-        sendForespoerselBesvart(oppdatertForespoerselId)
-
-        runBlocking {
-            val oppdatertFsp = hentForespoersel(oppdatertForespoerselId)
-            oppdatertFsp.navReferanseId shouldBe oppdatertForespoerselId
-            oppdatertFsp.status shouldBe Status.BESVART
-
-            val forespoersel = hentForespoersel(forespoerselId)
-            forespoersel.navReferanseId shouldBe forespoerselId
-            forespoersel.status shouldBe Status.FORKASTET
-        }
-    }
-
-    private fun sendForespoerselMottatt(
-        forespoerselId: UUID,
-        vedtaksperiodeId: UUID,
-    ) {
-        val mottattJson = buildForespoerselMottattJson(forespoerselId, vedtaksperiodeId)
-        Producer.sendMelding(ProducerRecord(priTopic, "key", mottattJson))
-    }
-
-    private fun sendForespoerselOppdatert(
-        oppdatertForespoerselId: UUID,
-        forespoerselId: UUID,
-        vedtaksperiodeId: UUID,
-    ) {
-        val oppdatertJson =
-            buildForespoerselOppdatertJson(
-                forespoerselId = oppdatertForespoerselId,
-                eksponertForespoerselId = forespoerselId,
-                vedtaksperiodeId = vedtaksperiodeId,
-            )
-        Producer.sendMelding(ProducerRecord(priTopic, "key", oppdatertJson))
-    }
-
-    private fun sendForespoerselBesvart(forespoerselId: UUID) {
-        val besvartJson = buildForspoerselBesvartMelding(forespoerselId)
-        Producer.sendMelding(ProducerRecord(priTopic, "key", besvartJson))
-    }
-
-    private suspend fun hentForespoersel(forespoerselId: UUID): Forespoersel {
-        val response =
-            fetchWithRetry(
-                url = "http://localhost:8080/v1/forespoersel/$forespoerselId",
-                token = mockOAuth2Server.gyldigSystembrukerAuthToken("810007842"),
-            )
-        return response.body()
-    }
-
-    private fun sjekkOmForespoerselFinnesIDB(
-        oppdatertForespoerselId: UUID,
-        eksponertForespoerselId: UUID?,
-        vedtaksperiodeId: UUID?,
-    ) {
-        runBlocking {
-            val responseOppdatertFsp =
-                fetchWithRetry(
-                    url = "http://localhost:8080/v1/forespoersel/$oppdatertForespoerselId",
-                    token = mockOAuth2Server.gyldigSystembrukerAuthToken("810007842"),
-                )
-            val oppdatertFsp = responseOppdatertFsp.body<Forespoersel>()
-            oppdatertFsp.navReferanseId shouldBe oppdatertForespoerselId
-            oppdatertFsp.status shouldBe Status.AKTIV
-            transaction(db) {
-                ForespoerselEntitet
-                    .selectAll()
-                    .where {
-                        (ForespoerselEntitet.navReferanseId eq oppdatertForespoerselId) and
-                            (ForespoerselEntitet.eksponertForespoerselId eq eksponertForespoerselId) and
-                            (ForespoerselEntitet.vedtaksperiodeId eq vedtaksperiodeId)
-                    }.count() shouldBe 1
-            }
-        }
-    }
-
-    private fun sjekkOmDetFinnesKunEnForespoerselIDB(forespoerselId: UUID?) {
-        runBlocking {
-            val response =
-                fetchWithRetry(
-                    url = "http://localhost:8080/v1/forespoersler",
-                    token = mockOAuth2Server.gyldigSystembrukerAuthToken("810007842"),
-                )
-            response.status.value shouldBe 200
-            val forespoerselSvar = response.body<List<Forespoersel>>()
-            forespoerselSvar.size shouldBe 1
-            forespoerselSvar.first().navReferanseId shouldBe forespoerselId
-        }
     }
 }
